Move tree workflow data into its own Vuex module to avoid confusion
